+++ localflavor/ch/forms.py
@@ -92,17 +92,17 @@ class CHIdentityCardNumberField(Field):
 
         match = re.match(id_re, value)
         if not match:
-            raise ValidationError(self.error_messages['invalid'])
+            raise ValidationError(self.error_messages['invalid'], code='invalid')
 
         result = match.groupdict()
         idnumber, pos9, checksum = result['idnumber'], result['pos9'], result['checksum']
 
         if idnumber in ('00000000', 'A0000000'):
-            raise ValidationError(self.error_messages['invalid'])
+            raise ValidationError(self.error_messages['invalid'], code='invalid')
 
         all_digits = "%s%s%s" % (idnumber, pos9, checksum)
         if not self.has_valid_checksum(all_digits):
-            raise ValidationError(self.error_messages['invalid'])
+            raise ValidationError(self.error_messages['invalid'], code='invalid')
 
         return '%s%s%s' % (idnumber, pos9, checksum)
 
@@ -135,4 +135,4 @@ class CHSocialSecurityNumberField(CharField):
             super().run_validators(value)
         except ValidationError as errs:
             # Deduplicate error messages, if any
-            raise ValidationError(list(set(errs.messages)))
+            raise ValidationError(list(set(errs.messages)), code='invalid')